Fix rare cases of using uninitialized values in LMR
This was obviously a dumb design decision to begin with. As far as I know 1.1.48 (#62) has caused the issue to start appearing, however, it went largely unnoticed due to matching benches.
Eventually, a recent case of diverging benches on a non-main branch prompted the investigation into this.

Renegade dev 1.2.10
Bench: 3906701